Modified Id Datatype from Long to Integer

+++ src/main/java/com/example/FilmBuffs/dto/request/AddReviewRequest.java
@@ -13,7 +13,7 @@ import lombok.*;
 public class AddReviewRequest {
 
     @NotNull
-    private Long filmId;
+    private Integer filmId;
 
     @NotBlank
     private String review;
